chore(storybook config): reorg storybook configs to live on root
affects: @crave/farmblocks-alert, @crave/farmblocks-tooltip

chore(snapshots): remove __snapshots__ folder from individual packages and update the main one in th

affects: @crave/farmblocks-alert, @crave/farmblocks-tooltip

refactor(storybook): include iconfont and global css on the root storybook config

affects: @crave/farmblocks-button
