--- packages/alert/.storybook/config.js
@@ -1,9 +1,0 @@
-import { configure } from "@storybook/react";
-
-const req = require.context("../src/", true, /\.story\.js$/);
-
-function loadStories() {
-  req.keys().forEach(filename => req(filename));
-}
-
-configure(loadStories, module);
