--- packages/tooltip/.storybook/addons.js
@@ -1,2 +1,0 @@
-import "@storybook/addon-actions/register";
-import "@storybook/addon-links/register";
